GBE-679: extract add_event from remains
